feat: extract controls
